Extract KeyInScopeResolver tests to separate files
