+++ resaca/src/main/java/com/sebaslogen/resaca/KeyInScopeResolver.kt
@@ -2,7 +2,6 @@ package com.sebaslogen.resaca
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.DisposableEffect
-import androidx.compose.runtime.remember
 import androidx.lifecycle.viewmodel.compose.viewModel
 
 /**
